QSDK:CP: hostapd: Add support for ATF offload distribution
Airtime Fairness (ATF) is a feature designed to maintain scheduling fairness
for traffic transmitted by the Access Point (AP). It regulates airtime
allocation to ensure reserved and controlled airtime is guaranteed
based on user requirements, enhancing network performance and efficiency.

This series brings initial infra of ATF and ATF distribution logic.

The following features are supported

Equal Distribution/Default behavior​ - when atf_offload is enabled without
any configuration.

ATF based on SSID​ - Configuration and Allocation of airtime to a specific SSID​.

ATF based on SSID Group​ - Configuration and Allocation of airtime which would be
		shared across all SSID’s in this group​.

ATF based on Peer  - Configuration and allocation of airtime to a specific
		peer-MAC Address based​ associated with group/ssid.

All the airtime configurations are based on percentage of 100 by user. But,
when it is calculated in hostapd it is scaled up by 10 i.e., 1000 to consider
one digit after the decimal point.

The distribution runs on the logical entity named as group. In case of ATF based
on ssid, each ssid will have one group and in case of group its default that group.
Each group will have configured and unconfigured peers which would be described
in detail below:

Airtime Distribution logic kicks in for following triggers
       1. commitatf command
       2. station join
       3. station leave.

Once any of the 3 events occur, a timeout is set for 1 second. After the timeout,
the distribution of airtime starts.

By default, each radio will have maximum airtime of 1000. This airtime will be
distributed among the group/ssid and peers.

Before starting the distribution, the algorithm will map all the ssid configs with
the actual bss that is operating and also, all the configured peers to the peers
authorized in that bss.

It considers all the peers and sorts them as implicit and explicit peers.

The user configured peers will be considered as explicit peers of the group and
unconfigured peers associated to that ssid (within the group) will be considered
as implicit peers.

Once the candidate list is ready. The algorithm starts distributing the airtime
for each configured group first.

for each configured group, it will calculate for implicit and explicit peers.

once all the configured group has completed the distribution, the residual airtime
of the radio will be shared among the default-group peers.

default-group is the group which contains all the unconfigured ssid of radio and
peers associated with those ssid.

Change-Id: I4138a09e93e290604a989f81ec46be5e09538877
Signed-off-by: Harshitha Prem <hprem@qti.qualcomm.com>