UPSTREAM:PROTOCOL:hostapd: Fix AWGN channel change failure
As per the recent changes made in the cfg80211, any 6 GHz channel change
request also needs to include the 6 GHz power mode as a mandatory
attribute.

As part of the AWGN channel change request, the 6 GHz power mode
is not included in the request, which results in the kernel rejecting
the channel change request.

To fix this, include the 6 GHz power mode in the AWGN channel change
request. Select the 6 GHz power mode as the best power mode for the
target channel.
Also, update other channel change requests to include the 6 GHz power
mode as well.

Since the channel change request already includes the 6 GHz power mode,
there is no need to store and track the CSA power mode separately.
So, remove the code that tracks the CSA power mode.

Make sure to not allow power mode change when the CSA is in progress.

Fixes: Ia9ef1bdef51 ("UPSTREAM: PROTOCOL: hostapd: Support 6 GHz SP mode with MLO")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
n00-001-hostapd-add-support-for-6GHz-operation.patch
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch
r66-UPSTREAM-PROTOCOL-hostapd-Handle-AFC-Channel-Change.patch
r90-UPSTREAM-PROTOCOL-hostapd-Support-6-GHz-SP-mode-with.patch

Change-Id: Ib75ab57da7a23ce17931bc79502d0d3dafb09b06
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>