UPSTREAM:PROTOCOL:hostapd: Apply dynamic puncturing only for SP mode
Hostapd computes and applies the best 6 GHz puncture pattern
for the interface during channel change. This dynamic puncturing
is applicable only for Standard Power (SP) mode where the EIRP
depends on the AFC payload. there is no usecase to apply dynamic
puncturing for other power modes.

Currently, even if the user has forced the power mode to LPI, the
dynamic puncturing is applied. This causes the interface to
operate in LPI mode with dynamic puncturing, which is not expected.

To fix this, make sure that the dynamic puncturing is applied
only when the power mode is set to Standard Power (SP) or if
the user has not provided any power mode configuration.

Fixes: Ie8f8d71ecf8 ("UPSTREAM: PROTOCOL: hostapd: Apply dynamic puncturing only for 6 GHz")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
n00-001-hostapd-add-support-for-6GHz-operation.patch
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch
r72-a-UPSTREAM-PROTOCOL-hostapd-6-GHz-power-mode-channels-.patch
r72-d-UPSTREAM-PROTOCOL-hostapd-Clean-up-and-refactor-6-GH.patch
r72-e-UPSTREAM-PROTOCOL-hostapd-Support-for-SP-and-Punctur.patch
r72-h-UPSTREAM-PROTOCOL-hostapd-Apply-best-power-mode-for-.patch
r80-d-UPSTREAM-PROTOCOL-hostapd-Add-minimum-PSD-calculatio.patch
r81-UPSTREAM-PROTOCOL-hostapd-Use-OOBE-compliant-PSD-for.patch

Change-Id: I0507b07326aed32894753554cb4284d78857906c
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>