UPSTREAM:PROTOCOL: hostap: free wpa_auth sm from previous failed authentication
Currently, when the initial Multi-Link (ML) authentication fails, the
wpa_auth state machine is not properly freed, leading to a memory leak.
This typically occurs when an ML STA begins authentication on one link but
roams to a partner link before reaching the authorized state. During this
roaming process, the link ID is updated, but the original wpa_auth state
machine is left behind. Since the station is not authorized, retaining the
state machine is unnecessary and results in a memory leak.

Add logic to free the wpa_auth state machine from partner links when their
previous ML authentication attempt has failed.

Patch Dependency: r36-02-Update-partner-linkId-for-proper-partner-link-reset.patch
Link: None
Feature: MLO

Signed-off-by: Aditya Kumar Singh <aditya.kumar.singh@oss.qualcomm.com>
Change-Id: I15cfcf8d8fc455e79dbc61fce1cf9fc89fa09a15