UPSTREAM:PROTOCOL: hostap: handle removal in progress flag or non-offloaded drivers
Driver offload multi-link removal introduced an in-progress flag to
determine whether a reconfiguration element should be added. However, this
flag is currently set only in the driver offload path and not in
non-offload drivers. As a result, the reconfiguration element is not added
in the non-offload case, leading to link removal failures

Add changes to handle the new flag in non-offload case as well.

Patches Dependency: r32-004-hostapd-Add-support-for-ML-reconfigure-offloaded-dri.patch
Link: None
Feature: Multi Link Removal

Fixes: a9b84d4802d5 ("hostapd: Add Multi-Link Reconfigure link removal support")
Signed-off-by: Aditya Kumar Singh <aditya.kumar.singh@oss.qualcomm.com>
Change-Id: Icf01c605cd075b793cc0dbe43fa6ccde3cd846a0