UPSTREAM:CP:wpa_supplicant: Fix missing update of 6G assoc freq
When the associated frequency is not set (i.e., 0) in the assoc_info structure,
`wpa_cli status` incorrectly reports a frequency of 0 for STA VAPs on 6GHz radios.
This occurs despite the BSSID being known and the frequency being available in
the BSS table.

This patch addresses the issue by checking if the frequency is zero and
`wpa_s->current_bss` is valid. If so, it directly assigns the frequency from
`current_bss->freq` to `assoc_info.freq`. This ensures that the correct
association frequency is reflected in the status output, particularly for
6GHz STA interfaces.

Change-Id: I886fdf6e39a5429caa0b337ba2a8a8f970cf34c6
Signed-off-by: Banka Sai Krishna <banksaik@qti.qualcomm.com>