QSDK:PROTOCOL:hostap: Add 240 MHz vendor element in (Re)association response
Currently, 240 MHz vendor element is getting added in association response
only for the assoc link. But, for non-assoc link, vendor element has to
be added in the per-sta profile as well. Because of this, station(s) are
not getting associated when 240MHz bandwidth is enabled in 5GHz link as
STA sees the bandwidth mismatch between (Re)association response and
beacon of the non-assoc link(s).

Fix this by adding the 240 MHz vendor element inside per-sta profile for
non-assoc link.

Feature: 240 MHz (QtoQ)

Change-Id: Id64d24596e0655e45054892a07b7918984d62e43
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>