UPSTREAM:PROTOCOL:hostapd:Correct PSD power assignment for straddle channels
When the AP operates on a channel that spans both LPI and SP power mode ranges
(e.g., channel 161 in HT320 under SP mode), certain sub-channels—such as
185 and 189—are not permitted in SP mode. These channels should therefore
use LPI PSD power levels.

However, it was observed that the system incorrectly assigns
the PSD power based on primary20MHz - 16, leading to improper power
configuration. This occurs because the current logic fails to recognize
that channels 185 and 189 are not SP-allowed, relying solely on the presence
of a channel pointer rather than checking the channel's
enabled/disabled state.

To address this, the fix introduces a validation step that checks
the disabled status of the channel instead of relying on the existence of
a channel pointer, ensuring accurate PSD power assignment for straddle channels.

Additionally, for punctured SP channels, the TPE should advertise
power as primary20MHzPSD - 16. The current implementation incorrectly
advertises primary20MHzPSD - 22. This patch fixes that as well.

Fixes:
3cca0f01c14(r85-00-UPSTREAM-PROTOCOL-hostapd-Handle-TPE-PSD-for-punctur.patch)
Patch work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency: r72-e-UPSTREAM-PROTOCOL-hostapd-Support-for-SP-and-Punctur.patch
n00-001-hostapd-add-support-for-6GHz-operation.patch
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch
r65-QSDK-HOSTAPD-Implement-Min-PSD-Calculation-and-Punct.patch
r72-a-UPSTREAM-PROTOCOL-hostapd-6-GHz-power-mode-channels-.patch
r72-d-UPSTREAM-PROTOCOL-hostapd-Clean-up-and-refactor-6-GH.patch
r85-00-UPSTREAM-PROTOCOL-hostapd-Handle-TPE-PSD-for-punctur.patch

Change-Id: I265304762f75d844fbe49e78da583784ec9818da
Signed-off-by: Priyadarshnee Srinivasan <priys@qti.qualcomm.com>