UPSTREAM:PROTOCOL:wifi:hostapd: Add RRB-based roam notification to reduce FT auth delays
When a client roams from a source AP to a target AP and returns to the
source with FT authentication, the AP first checks if the station is
already present in its station list. It deletes the station from all
associated links in hostapd down to the firmware, then creates a new
STA entry and sends the authentication reply.

Currently, the authentication reply is delayed by 400–450 ms.
The station retransmits two additional authentication requests
every 200 ms if it does not receive a reply. The AP processes
all three and sends consecutive authentication responses, updating
the Anonce each time. However, the STA sends a reassociation request
using the Anonce from the first response, while the AP expects the
latest Anonce, resulting in reassociation failure due to an invalid
FTIE.

This delay has increased following recent design changes after
a package upgrade.

To resolve this, introduce a new RRB message that sends a roam
notification to all APs in the same mobility domain once a station
successfully reassociates. Upon receiving this notification, other
APs delete the station entry if present. This ensures that when the
STA returns to the original AP, the authentication reply is sent
immediately, avoiding retries and reassociation failures.

Change-Id: Ibf331e3a8ea15d5f2740cde7322f5c54a5357c22
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>