UPSTREAM:PROTOCOL: hostapd: Mark radar channels as unavailable
Currently, when hostapd receives a radar detected event from the
driver with the radar bitmap, the current channel's puncture
bitmap is checked against the radar bitmap to determine
if the radar event is received for the punctured channel.

If the radar event is detected on a channel which is punctured,
the corresponding radar event from the driver is ignored as the
channel is punctured. But we do not mark the corresponding
channel's state as unavailable. This leads to the channel
remaining in the available state even though it is not usable
due to radar.

To fix this, mark the channel as unavailable when we receive
a radar event, even for a punctured channel. This ensures that
the channel is not used for any future operations until the
channel is cleared by the driver.

Also, the check to determine if the radar event is received
on a punctured channel only considers if one of the
channel in the radar bitmap is punctured. This is not correct
as the radar event can be received for multiple channels
and one or more of the sub-channel(s) may not be punctured.

To fix this, check if all the channels in the radar bitmap
are punctured before ignoring the radar event.

Fixes: I8b98685cee2 ("hostapd: add hostapd config option to enable dfs punturing")
Patch-work: None
Tag: DFS
Patch-dependency:
q02-055-hostapd-ru-puncturing-for-dfs-channels.patch

Change-Id: I60b986063ef505f24f2d011c18dcd0651ca4f383
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>