UPSTREAM:PROTOCOL: hostapd: Fix accessing removed bss during scan timeout
Currently if wifi down is given when vaps are trying to come up,
crash is seen this is because when scan is in progress and wifi down
is given, scan_timeout should have been cancelled by eloop_cancel_timeout
in wpa_driver_nl80211_deinit() but this cancel_timeout is not happening
for the bss->ctx for which the scan was initiated.

Scan would have been initiated with the first bss, but during
wpa_driver_nl80211_deinit() the bss which is removed last will be given
for eloop_cancel_timeout which does not match the one which initiates scan
and hence the scan_timeout is not cancelled for the bss which initiates the
scan which results in crash when the scan_timeout tries to access the bss
which got removed.

Fix by initiating eloop_cancel_timeout even before the links got removed
such that the scan_timeout would be cancelled properly and update the scan_link
with flink which got updated.

Patch-dependency: r65-UPSTREAM-PROTOCOL-hostapd-use-dormant_links-to-notif.patch

Patch-work: None

Fixes: I6f75d2f640fc(UPSTREAM:PROTOCOL: hostapd: use active_links to notify
start/stop state of links)

Change-Id: I288f1c8ea06103e2d9a2e6320f77b1d8bad0725d
Signed-off-by: Dhanavandhana Kannan <dhanavan@qti.qualcomm.com>
(cherry picked from commit f3eaf8a040265e83fd204e8e0b5ce66694c77b82)