UPSTREAM:PROTOCOL:hostapd: Fill sm->r1_key_holder even if PMK-R1 is not in local cache
During roaming, when the PMK-R1 key is not available in the local cache
(which is expected), the sm->r1_key_holder is not filled. As a result
AP reject reassocation request from client even reassociation request
is valid.

Ensure that sm->r1_key_holder is populated regardless of whether the
PMK-R1 key is present in the local cache.

Patch-dependency:
r21-hostapd-Add-FT-over-DS-support-as-per-D4.0-spec.patch
Patch-work: None

Fixes: bfa1c81a5be1 ("hostapd: Add FT over DS support as per D4.0 spec")

Change-Id: I048b20a49907b58925894a55feb8430604b0175f
Signed-off-by: Aaradhana Sahu <aaradhana.sahu@oss.qualcomm.com>