QSDK:PROTOCOL:hostapd: Update re-assoc mechanism frequency list
In split-phy case, in-order to identify whether the corressponding AP
has split-phy enabled, frequency list from wiphy info will be used.
For 5 GHz high band, start frequency was checked between 5490 and 5895.
With the latest upgrade, the start frequency was updated from 5490 to
5895. Because of this update, conditional logic is failing to enable
split-phy even if the AP is powered up in split-phy radio.

Fix this by updating the frequency logic to include both start and end
frequencies instead of checking the in-between range.

Patch-dependency: None
Patch-work: None

Change-Id: I57c2b96f849b4d87b15e545e5a1b7677d16f69d1
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>