QSDK:CP: hostapd: ATF: Handle reassoc scenario in ML station
When multi-link station connects and disconnects continuously,
it does reassociation because of this same sta's candidate list
memory (say X) is added twice to ATF peer list. When this peer
leaves only one entry is removed from the peer list, other node
becomes dangling pointer.

To handle this, if the station is reassociating remove the
old node and initialize the node, so that this can be reused
for peer addition during algorithm calculation. Also, intialize
the sta's candidate list during sta allocation.

Change-Id: I5c622cd1da5d6fdee95199f563371012b4939f45
Signed-off-by: Harshitha Prem <hprem@qti.qualcomm.com>