UPSTREAM:PROTOCOL:hostapd: Fix accessing freed pointer during iw del scenario
Currently when the iw del command is issued, the function
nl80211_stop_ap() is called for each link(say 2 GHz, 5 GHz,and 6 GHz
in a MLD). During this process, the active_links bitmap is updated,
and the flink  is reassigned to the next available active link after
each call.

For example:

nl80211_stop_ap() is called for 2 GHz → flink is updated to 5 GHz.
Then for 5 GHz → flink is updated to 6 GHz.
Finally for 6 GHz → all links are stopped, and active_links becomes zero.

After this, if a wifi down command is issued and links are removed in a
specific order (say 2 GHz first, then 6 GHz), the flink (which was pointing
to 6 GHz) gets removed. However, the driver context (drv->ctx) still points
to bss->ctx, which was associated with the now-deleted flink.

If an event arrives from the driver after this removal, it tries to
access drv->ctx, which now points to freed memory. This results in a
crash due to invalid memory access.

Fix by ensuring that flink is updated only during nl80211_remove_link()
and not during nl80211_stop_ap() the drv->ctx is properly updated with
flink->ctx during link removal and also ensure that scan_link is set to
NULL when the link being removed is same as the scan_link to prevent
accessing a freed pointer.

Patch-dependency:None

Patch-work: None

Fixes: I6f75d2f640fc(UPSTREAM:PROTOCOL: hostapd: use active_links to
notify start/stop state of links)
Fixes: I288f1c8ea061(UPSTREAM:PROTOCOL: hostapd: Fix accessing
removed bss during scan timeout)

Change-Id: Ib3be2d7de0f707cd2d0807e5a7c5908b47cfd634
Signed-off-by: Dhanavandhana Kannan <dhanavan@qti.qualcomm.com>