QSDK:PROTOCOL: Hostapd: Hidden BSSID Support for Tx and Non-Tx BSSes
Hidden BSSID profiles will be included in beacon and probe responses.
The SSID value will either be omitted or set to all zeros, based on
the configured value of ignore_broadcast_ssid.

Currently, Few issues have been identified where the SSID is not
included in unicast probe responses for requests targeting a hidden,
non-transmitted BSS, and where sending a probe request to a
non-transmitted BSS may unintentionally expose the SSID if the
transmitted BSS is hidden.

Add support to include the SSID in unicast probe responses for such
requests, depending on the ignore_broadcast_ssid configuration for
transmitted or non-transmitted BSS profiles

Fixes: 27faa86e745db (hostapd: QSDK patches update for package upgrade 13.0)

Change-Id: I129958c65cc728f6b848d7460608a371e2695e04
Co-developed-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>
Signed-off-by: Ramasamy Kaliappan <rkaliapp@qti.qualcomm.com>