UPSTREAM:PROTOCOL: hostapd: Fix AFC channel, bandwidth validation
Currently, during channel change to an SP channel, hostapd validates if
the channel and bandwidth configured is valid for the SP power mode.
However, this validation incorrectly takes the channel, BW and
puncture pattern from the current channel, instead of the target channel.

This results in an incorrect validation, where we end up doing a
channel change to a channel in SP mode, even though the target
channel parameters may not be valid for SP mode.

Fix this by using the input frequency, bandwidth and puncture
pattern to validate if the channel params are allowed for a
given power mode.

Fixes: I1a47bc82031 ("UPSTREAM:PROTOCOL:hostapd: Validate channel and bandwidth  during power mode configuration")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
n00-001-hostapd-add-support-for-6GHz-operation.patch
r28-hostapd-Add-hostapd-cli-support-to-switch-6GHZ-power.patch
r32-001-hostapd-Add-support-to-receive-AFC-power-update-comp.patch
r65-QSDK-HOSTAPD-Implement-Min-PSD-Calculation-and-Punct.patch
r72-a-UPSTREAM-PROTOCOL-hostapd-6-GHz-power-mode-channels-.patch
r72-d-UPSTREAM-PROTOCOL-hostapd-Clean-up-and-refactor-6-GH.patch
r72-e-UPSTREAM-PROTOCOL-hostapd-Support-for-SP-and-Punctur.patch
r96-00-UPSTREAM-PROTOCOL-hostapd-Validate-channel-and-bandw.patch

Change-Id: I97de942b10cb129cf12fca382ce6935f2a1f20e2
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>