QSDK:CP: wpa_supplicant.uc: Add changes to set radio mask irrespective of interface status
Currently, in wpa_supplicant.uc 'iface_start()', radio_mask is set only
when interface is not running already. This is to avoid starting the
interface again from ucode as well. But, this is not needed for
radio_mask as it can be set irrespective of that status.

Hence, add change to allow setting the radio_mask irrespective of the
interface status.

Change-Id: I20152dc4166df7cae6923278cf8cccf7ca1f9951
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>