QSDK:HOSTAPD: Remove r19-Fill-Supported-Op-Classes-IE-based-on-country.patch
The patch r19-Fill-Supported-Op-Classes-IE-based-on-country.patch was
added to QSDK to address an issue, where the opclass 117 was getting
displayed on the AP operating on channel 40 in 40 MHz bandwidth.
The global opclass 117 is absent in China operating class table as per
the IEEE80211 spec. Hence country based opclass table was used in the
patch.

But the opclass 117 being absent in China operating class table, seems
to be a gap in the IEEE80211 spec. Ideally an opclass value should be
present for 40 MHz channels.

Therefore, remove the changes added by
r19-Fill-Supported-Op-Classes-IE-based-on-country.patch from QSDK,
to address this spec gap and make the AP behave like before.

Tag: REGULATORY

Change-Id: Iad780e5e0d5e359949c335f844bfa2f345834572
Signed-off-by: Hariharan Basuthkar <hbasuthk@qti.qualcomm.com>