QSDK:CP hostapd:ATF:Fix: showatftable not displaying all connected stations
The showatftable command is previously limited to showing only ~40
clients due to a fixed buffer size of 4096 bytes. This is causing
incomplete output when more stations are connected. Updated the
implementation to log the station information directly to logread
instead of returning buffer, ensuring all connected stations are now
visible.

Change-Id: I75b55563167e37ccf95a8061fafa455827833223
Signed-off-by: Harish Rachakonda <rachakon@qti.qualcomm.com>