UPSTREAM:PROTOCOL:hostapd: Fix failure to disable FILS/UBPR
When dual-band is enabled, hostapd disables FILS/UBPR for 6GHz by
sending a value of 0 for FILS max interval or UBPR interval. But
the commands fails as 0 is incorrectly treated as an invalid
interval value.

Fix the issue by allowing 0 as a valid FILS/UBPR interval value.

Fixes: None

Patch-depdendency: None

Patch-work: None

Change-Id: I5b349ffd44aad7cfa5013061054c7f5dae32c126
Signed-off-by: Suraj P Kizhakkethil <suraj.kizhakkethil@oss.qualcomm.com>