UPSTREAM:PROTOCOL:hostapd:Fix multiple ANonce generation  during repeated FT auth
During roaming, if the AP delays its authentication response,
the station may retry the authentication request. The AP
responds to each retry, generating a new ANonce for every
authentication reply. However, the station uses the ANonce
from the first response to construct the FTIE in its
reassociation request. This mismatch causes the AP to reject
the reassociation due to an invalid FTIE.

This patch fixes the issue by ensuring that the AP generates
the ANonce only once per FT authentication attempt and
reuses the same ANonce for all authentication replies to the
same station. This ensures consistency in the FT handshake
and prevents reassociation failures.

Change-Id: Ia3eb729dcacea90977f2af09f4dfb6a655efd968
Signed-off-by: Shiva Sankar Gajula <sgajula@qti.qualcomm.com>
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>