UPSTREAM:PROTOCOL:hostapd:Set sae_pwe to SAE_PWE_BOTH in WPS network block
As Per IEEE80211 be section 12.12.9 Security constraints for EHT
When an EHT STA performs SAE authentication with a peer EHT STA,
it shall set the SAE Hash-to-element field to 1 in the RSNXE and
use the SAE hash-to-element method when SAE authentication is
negotiated.

12.4.4.2.3 Hash-to-element generation of the password element
with ECC groups
An SAE initiator that has identified a peer that supports this
method through the receipt of beacon/probe responses, shall derive
a secret element PT according to the following method and indicate
this by setting status code in SAE commit message as SAE_HASH_TO_ELEMENT.

For MLO operation, AP expects the SAE Commit message from station
with status code: WLAN_STATUS_SAE_HASH_TO_ELEMENT, without which the
AP would treat the station as Non-MLO and reply with an SAE commit
frame without MLIE. For supplicant to send SAE commit frame with
status code WLAN_STATUS_SAE_HASH_TO_ELEMENT, it checks for the
presence of H2E bit in RSNXE as well as its own capability, ie,
sae_pwe as 1/2 in the WPS network block.

For the WPS network profile to be able to connect both with a MLO AP
supporting hash to element method as well as with a non-MLO legacy
AP that supports hunt and peck mechanism only and does not support
hash to element (H2E bit in RSNXE=0), set sae_pwe to 2(supports both
has to element and hunt and peck methods) in the WPS network profile.

Tag: WPS
Patch-work: None
Patch-dependency: None

Change-Id: I5a04f7864f02525f1d9ae70bf7721e918a9191d0
Signed-off-by: Sai Pratyusha Magam <smagam@qti.qualcomm.com>