QSDK:CP ATF:Add configured airtime to showatftable
The `showatftable` output previously not printing configured
airtime. This change fixes it by populating the
configured airtime field correctly.

Currently when ATF stats is enabled and `atf_stats_timeout` is not
explicitly configured, `g_atf_stats_timeout` incorrectly shows as 0.
This change sets the default timeout to 30 seconds when ATF stats are
enabled.

Change-Id: I7b671c616b377eb94ab6c30995eeb1fe2070dc05
Signed-off-by: Harish Rachakonda <rachakon@qti.qualcomm.com>