UPSTREAM:PROTOCOL:hostapd: Support MLD STA reconnection on same band
In the current design, when an authentication or association request
is received, the mac80211 driver translates the link address to the
MLD address and forwards it to hostapd. Hostapd processes the MLD address
and replies using the same. The ath12k driver then translates the
MLD address back to the link address.

Initially, an MLD AP (with 3 links) and an MLD STA are connected using
2 links. Later,without deauthenticaion, the same authorized MLD STA
attempts to connect as a legacy STA or MLD STA using the same MLD
address on the same band. It sends a auth/assoc request in the mld address,
but the MLD AP responds using the previous STA's link address, causing
connectivity issues.

Add changes to avoid address translation between link and MLD
addresses in the RX/TX path for auth/assoc frames.

Hostapd receives the link address directly in auth/assoc frames.
It find the mld address using MLD IEs and partner link information,
and responds using the correct link address.

Change-Id: I70c6bde4a981fe21b8a85efceaf29c705e3472ec
Signed-off-by: Ramasamy Kaliappan <rkaliapp@qti.qualcomm.com>
Co-developed-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>