UPSTREAM:PROTOCOL:Hostapd: Fix CCA stall issue during collision handling
When a user-initiated CCA is in progress and the color collision handler
timeout expires, hostapd assigns a new random color, overwriting the
settings of the ongoing user-initiated CCA operation. Additionally,
the CCA element in the beacon for the user initiated session may be
erroneously updated by the new CCA session before it is removed by
the CCA complete event from the user initiated session. This leads
to a failure in generating the CCA complete event for the subsequent
session, causing hostapd to reject any further color_change commands
due to the previous state machine not being properly cleared.

Fix this by avoiding registration of the color collision handler or
the initiation of a user-initiated session if a CCA is in progress
for any BSS in the interface. Also, cancel any existing timeout handler
when a user-initiated CCA process starts.

Change-Id: I6a0ecd514d7bbb3fe8453319de247ab0b12d79bb
Signed-off-by: Suraj P Kizhakkethil <suraj.kizhakkethil@oss.qualcomm.com>