UPSTREAM:PROTOCOL:hostapd: Skip pruning of unauthorized STA in partner link
In scenarios where an MLD AP receives authentication/association requests
on link0 from a non mld station, the connection may fail, resulting
in a STA entry that remains unauthorized.
Later, when the same station connects successfully to link1 of
the same MLD AP after completing 4-way handshake, the current design
checks for the STA in the partner link and prunes the association in
the partner hapd.

However, the current logic does not skip pruning when the STA is
unauthorized in the partner link, leading stale STA entry in partner hapd.

Adds a fix to skip pruning and STA removal in the partner hapd when
the STA is unauthorized.

Patch-dependency: NA
Patch-work: NA

Change-Id: Ifd283059906b4ad6e458acfdd64606203ff04c25
Signed-off-by: Ramasamy Kaliappan <rkaliapp@qti.qualcomm.com>