UPSTREAM:PROTOCOL:hostapd:Handle address translation for FT and FILS auth
During FT and FILS authentication, the destination address for the authentication
reply is fetched from the SM address, which corresponds to the MLD address of the
station. However, this MLD address is not the associated link address, causing
the station to miss the authentication reply.

Fix this by sending the authentication reply to the associated link address
stored in sta->reply_addr instead of the SM address.

Patch-dependency: s021-UPSTREAM-PROTOCOL-hostapd-Support-MLD-STA-reconnecti.patch
Patch-work: NA

Change-Id: Ib6cab3014c49eb2c90794538d7ee04d2c9ffe796
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@qti.qualcomm.com>