UPSTREAM:PROTOCOL: hostapd: Fix removing disabled BSS
Currently, if AP link reconfigure is initiated on a disabled BSS, the
request is sent to driver which leads to indefinite behavior as the
corressponding vdev is already deleted/stopped.

Fix this by checking whether the BSS is in enabled state before
initiating the link reconfigure procedure.

Patch-dependency: r32-004-hostapd-Add-support-for-ML-reconfigure-offloaded-dri.patch
Patch-work: None

Fixes: a9b84d4802d5 ("hostapd: Add Multi-Link Reconfigure link removal support")
Change-Id: I97ac9de04c0c0b85a93e83922cbafca0e1481ca6
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>