QSDK:PROTOCOL: hostapd: Channel switch to same channel for DFS test mode
Currently, when a 5 GHz link is configured in DFS test mode for
DFS probability detection, the expectation is to send out CSA with
the next channel in the CSA IEs same channel as the current channel.

To accomplish this, the hostapd adds a fixed beacon template with
CSA count hardcoded as 3. This is done as a beacon template update.

The CSA IEs are removed by hostapd after a timeout of 1 sec and
a new beacon template is sent to the driver.

Since the CSA IE is added to a fixed template and not as part of
actual CSA, the CSA offsets are not filled when the beacon template
is sent to FW.

As part of a recent beacon template validation that got added in FW,
an error is seen as the template has CSA IEs added, but the CSA
offsets were not updated properly. Consequently, the FW suspends the
transmissions due to the failure. This causes DFS probability detection
tests to fail.

To fix this, instead of adding a fixed beacon template with CSA IE,
trigger a channel switch to the same channel. The CSA offsets will be
computed correctly as computed for a regular channel switch.

For channel switch, use a CSA count of 5 which is used for other
hostapd initiated channel switches instead of current value of 3
for consistency.

Tag: DFS
Patch-dependency:
a00-004-Add-provision-to-test-RADAR-detection-probablity-in-.patch
Fixes: I3d68b12ec8b ("hostapd: Add provision to test RADAR detection probablity in dfs testing")

Change-Id: I3fac583d9b58fa2f66caffbde896a403b3dc99fe
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>